v2/pyproject.toml
[tool.poetry]
name = "sc-flow"
version = "0.1.0"
description = ""
authors = ["John Dodson <john.dodson@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.53"
langchain = "^0.3.14"
langchain-community = "^0.3.14"
langchain-experimental = "^0.3.4"
pytest = "^8.3.4"
environs = "^14.1.0"
pypdf = "^5.1.0"
uvicorn = "^0.34.0"
azure-ai-ml = "^1.23.1"
azure-storage-blob = "^12.24.0"
azure-identity = "^1.19.0"
neo4j = "^5.27.0"
llmsherpa = "^0.1.4"
graphdatascience = "^1.13"
tiktoken = "^0.8.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
copilotkit = "0.1.34"
setuptools = "^75.8.0"
motor = "^3.6.1"
python-multipart = "^0.0.20"
azure-search-documents = "^11.5.2"
mlflow = "^2.20.0"
sqlmodel = "^0.0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scflow = "sc_flow.scflow:main"
